Create Team.java
The Team class represents a team in the system and extends the Entity class. Each team contains a list of players. The addPlayer() method allows for adding a player to the team by name. If a player with the same name already exists, the existing player is returned; otherwise, a new player is created and added to the team. This class also provides a formatted string representation of the team's id and name.